moved the socket middleware in middleware folder
